backend/requirements.txt
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-graph-stores-neo4j
python-dotenv
fastapi
uvicorn[standard]
nest-asyncio

# --- Smart Chunking / Layout Analysis / Cleaning ---
pdfplumber          # Layout-aware PDF extraction + table detection
ftfy                # Fix broken Unicode / mojibake encoding
spacy               # NLP for heuristic coreference resolution
python-docx         # DOCX layout-aware loading
beautifulsoup4      # HTML boilerplate stripping
lxml                # Parser backend for BeautifulSoup
easyocr             # Local OCR for images and scanned PDFs
Pillow              # Image handling for OCR

# --- BM25 Sparse Retrieval ---
llama-index-retrievers-bm25   # Lexical keyword search for hybrid retrieval

# --- Graph Cleaning & Community Detection ---
numpy               # Vectorized cosine similarity for graph cleaning
networkx            # Louvain/Leiden community detection for GraphRAG

# --- Research Engine ---
arxiv               # ArXiv API for paper discovery
rich                # Premium CLI interface / REPL
ddgs                # Free web and news searches
trafilatura         # Clean text extraction from web pages
